+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# jervis 1.7
+# jervis 1.7 - Apr 14th, 2020
 
 ### Bug fixes